+++ static/app/views/starfish/views/webServiceView/failureDetailPanel.tsx
